Moved LoadPlugin_sqlite3_plugin into SQLite3DB
In this way we can initialize it from other applications, for example TAP tests
