+++ test/tap/tests/sqlite3-t.cpp
@@ -12,7 +12,7 @@
 
 int main() {
 	plan(8);
-
+	SQLite3DB::LoadPlugin(NULL);
 	SQLite3DB *db;	// in memory
 	db = new SQLite3DB();
 	db->open((char *)"file:mem_db?mode=memory&cache=shared", SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE | SQLITE_OPEN_FULLMUTEX);
